Updated BTB_FSM
FSM is able to switch between states correctly, but still need to add in initial value states so that it can change based on the current state that the FSM doesn't know before hand.

This may require updating the FSM code so that a clock is not used or something. More thinking required